Se crea de consulta de incapacidad (PDS-4348)
