+++ src/test/resources/Pages/DEV/EvolutionWave/consultas/incapacidades/Incapacidades.json
@@ -1,0 +1,34 @@
+{
+  "Desde": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//mat-form-field[1]//div[1]//div[1]//div[4]"
+  },
+  "2020": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//div[contains(text(),'2020')]"
+  },
+  "Febrero": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//div[contains(text(),'FEB.')]"
+  },
+  "5": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//div[contains(text(),'5')]"
+  },
+  "Hasta": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//mat-form-field[2]//div[1]//div[1]//div[4]//mat-datepicker-toggle[1]//button[1]"
+  },
+  "10": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//div[contains(text(),'10')]"
+  },
+  "Consultar": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//button[@class='mat-focus-indicator mat-raised-button mat-button-base mat-primary']"
+  },
+  "Tabla": {
+    "GetFieldBy": "Xpath",
+    "ValueToFind": "//table[@class='mat-table cdk-table']"
+  }
+}
