--- src/test/resources/Pages/DEV/EvolutionWave/consultas/incapacidades/Texto dummy.txt
@@ -1,3 +1,0 @@
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur nec lacinia leo. In in elit tortor. Aliquam eget consectetur augue. Duis nibh nisi, pretium sit amet lobortis quis, vulputate non sapien. Suspendisse auctor diam at velit feugiat, vel accumsan tortor volutpat. Ut molestie semper metus id facilisis. Nunc at egestas neque, non finibus quam. Nulla metus enim, malesuada quis blandit eu, condimentum eget tortor.
-
-In est mauris, tincidunt a efficitur ut, bibendum a mi. Suspendisse potenti. Nullam iaculis magna odio, vel tempus orci tincidunt vel. Proin volutpat non sapien at lobortis. Sed eu posuere eros. Nam commodo interdum tristique. Nunc ullamcorper et lacus et faucibus. Ut in leo non diam blandit tempor. Sed volutpat, erat tincidunt iaculis scelerisque, lectus nisi varius orci, eu ultrices metus lacus vitae urna. Pellentesque vitae nisl lacus. Nulla facilisi. Quisque metus risus, accumsan eu nibh a, placerat tempor tellus. Cras sodales suscipit purus, sed sodales volutpat.
